[02/22 jyb] get and set user profile image works but causes errors

+++ web/flask-server/server.py
@@ -10,6 +10,9 @@ import pymysql
 from datetime import datetime
 import pandas as pd
 
+from PIL import Image
+import io
+
 from sqlalchemy import extract, asc
 from sqlalchemy.exc import IntegrityError
 
@@ -253,31 +256,30 @@ def get_current_user():
 		"createdAt": user.created_at
 	})
 
+
 @app.route("/get_profile_image", methods=['POST'])
 def get_user_profile_image():
-	username = request.json['username']
-	
-	# fetch user data by username(from session) from user_info_db : user_account table
-	user = User.query.filter_by(username=username).first()
-	
-	if user.profile_img is not None:
-		try:
-			image_data = bytearray()
+	try:
+		username = request.json['username']
+		
+		if not username:
+			return jsonify({"error": "Username not provided"}), 400
 			
-			ssh_client.connect(ssh_host, port=ssh_port, username=ssh_username, password=ssh_password)
-			sftp = ssh_client.open_sftp()
-			
-			with ssh_client.open_sftp() as sftp:
-				file = sftp.file(user.profile_img)
-				image_data.extend(file.read())
-			sftp.close()
-			ssh_client.close()
-			
-			return send_file(bytearray(image_data), mimetype='image/jpeg')
-		except Exception as e:
-			return str(e), 500
-	else:
-		return "No image", 404
+		# fetch user data by username(from session) from user_info_db : user_account table
+		user = User.query.filter_by(username=username).first()
+		
+		if not user or not user.profile_img:
+			return jsonify({"error": "Image not found"}), 404
+		
+		ssh_client.connect(ssh_host, port=ssh_port, username=ssh_username, password=ssh_password)
+		with ssh_client.open_sftp() as sftp:
+			with sftp.file(user.profile_img, 'rb') as file:
+				image_data = file.read()
+				return send_file(io.BytesIO(image_data), mimetype='image/png')
+		
+	except Exception as e:
+		print(str(e))
+		return jsonify({"error": "Internal server error"}), 500
 	
 
 @app.route("/set_profile_image", methods=['POST'])
@@ -286,24 +288,31 @@ def set_profile_image():
 	
 	# fetch user data by username(from session) from user_info_db : user_account table
 	user = User.query.filter_by(username=user_id).first()
-	if user:
+	print(request.files['image'])
+	
+	if (user and ('image' in request.files)):
 		try:
 			image_file = request.files['image']
-			# convert image_file format to jpg if needed
-			local_image_path = 'web/temp/'+user_id+'.jpg'
+			# Save image locally (temporarily)
+			local_image_path = 'temp/image.jpg'
 			image_file.save(local_image_path)
-			remote_image_path = 'D:/log/profile/'+user_id+'.jpg'
+			
+			remote_image_path = 'D:/log/user/' + user_id + '.jpg'
 			
 			ssh_client.connect(ssh_host, port=ssh_port, username=ssh_username, password=ssh_password)
 
 			with ssh_client.open_sftp() as sftp:
 				sftp.put(local_image_path, remote_image_path)
-			
+
 			os.remove(local_image_path)
-			
+		
 			ssh_client.close()
-			
+		
+			user.profile_img = remote_image_path
+			db.session.commit()
+		
 			return 'Successfully added profile image!', 200
+		
 		except Exception as e:
 			print(f"Error in record: {str(e)}")
 			return 'Error setting profile image', 500
